+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # itk-os2display/campaign-bundle
 
+## in develop
+
+* Made save button sticky.
+* Fixed error messages.
+
 ## 1.0.0
 
 * First release.